Improve VMSQL behaviour
Create database_storage_mode shared constant which can be set to :vmsql by subclasses to enable vmsql behaviour.
Add vmsql_additional_dataset_args so that it is declared in this central location.
Change create_database to add the vmsql to any geometric dataset that are being created.  This is needed because there are test methods which hard-code geometric dataset definitions locally and then simply calls create_database() with it.